Enable to resume

+++ src/main.rs
@@ -23,6 +23,10 @@ async fn main() -> Result<(), Box<dyn std::error::Error>> {
         let id = csv_data.id;
         let url = csv_data.image_url;
         let extension = Path::new(&url).extension().and_then(OsStr::to_str).unwrap();
+        if Path::new(format!("{save_path}{id}.{extension}").as_str()).exists() {
+            println!("{id}: おもわずスキップス♪");
+            continue;
+        }
         let image_bytes = reqwest::get(&url).await?.bytes().await?;
         let mut saving_file = File::create(format!("{save_path}{id}.{extension}"))?;
         std::io::copy(&mut image_bytes.as_ref(), &mut saving_file)?;
